KL-177/style: separate constant datas with files
- 반복되는 상수 데이터를 파일로 분리하였습니다.
