+++ src/constants/initialPageData.js
@@ -1,0 +1,7 @@
+const initialPageData = {
+  page: 0,
+  size: 9,
+  sort: ['string'],
+}
+
+export default initialPageData
