+++ src/hooks/useProductData.js
@@ -1,14 +1,11 @@
 import { useState } from 'react'
+import initialPageData from '../constants/initialPageData'
 import useProductQuery from './useProductQuery'
 import useKyQuery from './useKyQuery'
 import parseQueryParams from '../utils/parseQueryParams'
 
 function useProductData() {
-  const [pageData, setPageData] = useState({
-    page: 0,
-    size: 1,
-    sort: ['string'],
-  })
+  const [pageData, setPageData] = useState(initialPageData)
   const selectedQueryArray = useProductQuery(pageData.page, setPageData)
   const uri = parseQueryParams('products', {
     ...pageData,
